Former-commit-id: 629cd4de9716b3ee37946c3add1dbde7444a620b

+++ modules/SofaBaseLinearSolver/MatrixLinearSolver.h
@@ -308,8 +308,10 @@ public :
 
     int getNbGroups() const
     {
-        if (isMultiGroup()) return this->groups.size();
-        else return 1;
+        if (isMultiGroup())
+            return (int)this->groups.size();
+        else
+            return 1;
     }
 
     void setGroup(int i)
